Reworked the infrastructure for validation rules and inference rules (#64)
Main motivation for this PR is to improve the performance of validation
and inference rules:
- Do not apply each validation/inference rule to each language node,
since most rules are related only to few language nodes.
- Therefore register these rules to language keys, calculate the
language key of the current language node to validate/infer and apply
only those rules, which are registered to this language key (inspired by
the ValidationRegistry of Langium!)
- `undefined` remains as a fall-back solution
- Another benefit is, that validation _and_ inference rules can be
defined in the same way, as you would register validation checks in
Langium! See OX and LOX for examples.

Beyond that, this PR contributes:
- New concept of "language keys" of language nodes to register inference
and validation rules, with the new service `LanguageService`
- Nearly all TypeScript types in the source code have `<LanguageType>` in order to get rid of `unknown` in the source code. Benefit:
In `typir-langium` projects, we work everywhere with `AstNode` (instead
of `unknown`). In our internal test projects, we have now
`TestLanguageNode` instead of `unknown`. Downside: The additional
generics in the code make reading/writing the source code slower.
- The factory API to create new types (e.g.
`typir.factory.Primitives.create({ ... })`) is extended with a chaining
API to register inference rules which are dedicated for the currently
created type (e.g.
`typir.factory.Primitives.create({...}).inferenceRule({...}).finish()`):
This makes the definition of inference rules more (TypeScript-)type-safe
and allows to define an arbitrary number of these rules.
- Create the predefined validations using the factory API, e.g.
`typir.factory.Functions.createUniqueFunctionValidation()`
- Reworked the API of validation rules to create validation hints:
Instead of returning `ValidationProblem`s, they need to be given to the
`ValidationProblemAcceptor` now, which is provided as additional
argument inside validation rules (strongly inspired by Langium!)
- Lots of fixed bugs, refactorings and new utilities
- Test utilities for expected validation issues